+++ src/App.js
@@ -7,21 +7,21 @@ import Login from "./pages/Login";
 import BookReservation from "./pages/BookReservation";
 import ContactUs from "./pages/ContactUs";
 import Events from "./pages/Events";
-import AddFoodForm from "./components/Body/AddFoodForm";
+import AddFoodForm from "./components/AdminControls/AddFoodForm";
 import Signup from "./pages/Signup";
 
 
 import {selectIsAdmin, selectIsVerifiedAdmin, selectUser} from "./store/features/userSlice"
 import { useSelector } from "react-redux";
 import { useEffect, useState } from "react";
-import AllMenu from "./components/Body/AllMenu";
+import AllMenu from "./components/MenuList/AllMenu.js";
 import Cart from "./pages/Cart";
 import Checkout from "./pages/Checkout";
 import useVerifyAdmin from "./Hook/useVerifyAdmin";
 import { Nav } from "react-bootstrap";
 import AdminDashBoard from "./pages/AdminDashBoard";
 import NotAuthorized from "./pages/NotAuthorized";
-import ViewUsers from "./components/Body/ViewUsers";
+import ViewUsers from "./components/AdminControls/ViewUsers";
 import ForgotPassword from "./pages/forgotPassword";
 import PasswordReset from "./pages/PasswordReset";
 import Error from "./pages/Error";
