+++ src/components/HomePage/Welcome.js
