+++ src/components/MenuList/DrinksMenu/Beverages.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import useFetchDrinks from '../../Hook/useFetchDrinks';
+import useFetchDrinks from '../../../Hook/useFetchDrinks';
 import { useDispatch, useSelector } from 'react-redux';
-import { AddToCartArray } from '../../store/features/cartSlice';
+import { AddToCartArray } from '../../../store/features/cartSlice';
 import { Link } from 'react-router-dom';
-import useHandleDelete from '../../Hook/useHandleDelete';
-import { selectIsAdmin } from '../../store/features/userSlice';
+import useHandleDelete from '../../../Hook/useHandleDelete';
+import { selectIsAdmin } from '../../../store/features/userSlice';
 
 
 function Beverages() {
