+++ src/components/MenuList/DrinksMenu/CoffeeAndTea.js
@@ -1,30 +1,29 @@
 import React from 'react'
-import useFetchDrinks from '../../Hook/useFetchDrinks';
+import useFetchDrinks from '../../../Hook/useFetchDrinks';
 import { useDispatch, useSelector } from 'react-redux';
-import { AddToCartArray } from '../../store/features/cartSlice';
+import { AddToCartArray } from '../../../store/features/cartSlice';
 import { Link } from 'react-router-dom';
-import useHandleDelete from '../../Hook/useHandleDelete';
-import { selectIsAdmin } from '../../store/features/userSlice';
+import useHandleDelete from '../../../Hook/useHandleDelete';
+import { selectIsAdmin } from '../../../store/features/userSlice';
 
-function SoftDrinks() {
+function CoffeeAndTea() {
 
   const isAdmin=useSelector(selectIsAdmin);
 
   const dispatch=useDispatch();
+  const {data,message,isLoading}=useFetchDrinks("Coffee & Tea");
 
-  
+  const {foodtype,handleDelete}=useHandleDelete(data);
 
-    const {data,message,isLoading}=useFetchDrinks("Soft Drinks");
-
-    const {foodtype,handleDelete}=useHandleDelete(data);
 
   return (
-  <div className="softdrinks">
-     <div className="container">
-        <h1 className="text-white bg-success title-menu" style={{padding:"10px",borderRadius:"20px"}}>Soft Drinks</h1>
+  <div className="coffee-tea">
+    <div className="container">
+        <h1 className="text-white bg-success title-menu" style={{padding:"10px",borderRadius:"20px"}}>Coffee & Tea</h1>
       </div>
-    
-          <div className="container border-container">
+
+
+      <div className="container border-container">
             {foodtype &&
               foodtype.map((drink) => (
                 <div className="drink-items" key={drink.id}>
@@ -45,8 +44,9 @@ function SoftDrinks() {
                 </div>
               ))}
               </div>
+
   </div>
   )
 }
 
-export default SoftDrinks
+export default CoffeeAndTea
