+++ src/pages/Home.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import Navbar from "../components/Header/Navbar";
-import Welcome from "../components/Body/Welcome";
-import Foods from "../components/Body/Foods";
-import About from "../components/Body/About";
-import Ratings from "../components/Body/Ratings";
+import Welcome from "../components/HomePage/Welcome";
+import Foods from "../components/HomePage/Foods";
+import About from "../components/HomePage/About";
+import Ratings from "../components/HomePage/Ratings";
 import Footer from "../components/Footer/Footer";
 
 function Home() {
