+++ src/store/reduxStore/store.js
@@ -11,7 +11,7 @@
 
 // reduxStore/store.js
 
-import { useSyncExternalStore } from 'react-redux';
+
 
 import { configureStore } from "@reduxjs/toolkit";
 import { persistReducer, persistStore } from "redux-persist";
